Rename MapLegend to MapSelectorBar & rudimentary Selector Bar functionality
Reduce misunderstandings
