+++ src/apps/Map/components/MapContent.js
@@ -2,24 +2,19 @@ import React from 'react';
 import { Route } from 'react-router-dom';
 import styled from 'styled-components';
 
-import MapLegend from '~/apps/Map/components/MapLegend';
+import MapSelectorBar from '~/apps/Map/components/MapSelectorBar';
 import MapSwitch from '~/apps/Map/components/MapSwitch';
 import config from '~/config';
-import { orientation } from '~/styles/utils';
-
-function getHeight(props) {
-  if (props.isEmbedMode) {
-    return 'auto';
-  }
-
-  return props.hasLegend ? '180px' : 0;
-}
+import { orientation, media } from '~/styles/utils';
 
 const MapContent = styled.div`
   width: 100%;
   background: #fff;
   flex: 0 1 auto;
-  min-height: ${getHeight};
+
+  ${media.m`
+    display: flex;
+  `}
 
   ${orientation.landscape`
     min-height: 80px;
@@ -40,7 +35,7 @@ export default (props) => (
           path={config.routes.map.hbiIndex}
           render={() =>
             props.displayLegend && (
-              <MapLegend
+              <MapSelectorBar
                 type="hbi"
                 filterHbi={props.filterHbi}
                 isEmbedMode={props.isEmbedMode}
@@ -53,7 +48,10 @@ export default (props) => (
           path={config.routes.map.projectsIndex}
           render={() =>
             props.displayLegend && (
-              <MapLegend type="plannings" isEmbedMode={props.isEmbedMode} />
+              <MapSelectorBar
+                type="plannings"
+                isEmbedMode={props.isEmbedMode}
+              />
             )
           }
         />
