+++ src/apps/Map/components/MapSelectorBar/SelectorBarImageWrapper.js
