+++ src/apps/Map/components/MapSelectorBar/SelectorBarItem.js
