+++ src/apps/Map/components/MapSelectorBar/SelectorBarWrapper.js
