+++ src/apps/Map/components/MapSelectorBar/images/hbi-ok@2x.png
