+++ src/apps/Map/components/MapSelectorBar/images/hbi-super@2x.png
